Switch default to Python3, and change deps installation
`setup.sh` used to install some dependencies with `apt-get` and some
with `pip`. In the pip part, we had `livestreamer`, which is not inuse,
and `youtube-dl` and `bottle` which both exist in Raspbian repositories
for Python3. So instead of the mix (which is problematic, because no one
ever updates pip deps), we install it all with apt-get.

This required to switch to Python3. The problem is that existing users
with a self-updating git installation will not see this change. So for
backwards compatibility I added a simple dependency check before running
the main script.